add react-router-dom and navigation
