+++ public/index.html
@@ -24,6 +24,17 @@
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
 
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+    <style>
+      * {
+        margin: 0;
+      }
+    </style>
+    <style>
+      .link {
+        text-decoration: none;
+        color: inherit;
+      }
+    </style>
 
     <title>React-Blog</title>
   </head>
