--- src/components/topbar/TopBar.js
@@ -1,33 +1,0 @@
-import "./topbar.css";
-
-const TopBar = () => {
-  return (
-    <div className="top">
-      <div className="topLeft">
-        <i className="topIcon fa-brands fa-facebook-square"></i>
-        <i className="topIcon fa-brands fa-twitter-square"></i>
-        <i className="topIcon fa-brands fa-pinterest-square"></i>
-        <i className="topIcon fa-brands fa-instagram-square"></i>
-      </div>
-      <div className="topCenter">
-        <ul className="topList">
-          <li className="topListItems">HOME</li>
-          <li className="topListItems">ABOUT</li>
-          <li className="topListItems">CONTACT</li>
-          <li className="topListItems">WRITE</li>
-          <li className="topListItems">LOGOUT</li>
-        </ul>
-      </div>
-      <div className="topRight">
-        <img
-          className="topImg"
-          src="https://images.unsplash.com/photo-1494790108377-be9c29b29330?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=2787&q=80"
-          alt="profileimg"
-        />
-        <i className="topSearchIcon fa-solid fa-magnifying-glass"></i>
-      </div>
-    </div>
-  );
-};
-
-export default TopBar;
